chore: restructured components and extracted nested components inside places auto complete component
