+++ .eslintrc.js
@@ -39,7 +39,8 @@ module.exports = {
     ],
     'quote-props': 0,
     'react/jsx-one-expression-per-line': 0,
-    'class-methods-use-this': 0
+    'class-methods-use-this': 0,
+    'object-curly-newline': ["error", { "consistent": true }]
   },
   overrides: [
     {
