+++ src/App.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import Routes from './routes';
 
 import Navbar from './components/Navbar/Navbar';
