+++ src/components/PropertySearch/PropertyInformation/PropertyInformation.scss
