+++ src/routes.js
@@ -1,5 +1,5 @@
 import { Route, Switch, withRouter } from 'react-router-dom';
-import React, { Component } from 'react';
+import React from 'react';
 
 import Homepage from './components/Homepage/Homepage';
 import SignUp from './components/SignUp/SignUp';
@@ -9,24 +9,16 @@ import InfoRequestPage from './components/InfoRequestPage/InfoRequestPage';
 import Profile from './components/Profile/Profile';
 import PropertySearch from './components/PropertySearch/PropertySearch';
 
-class Routes extends Component {
-  // constructor(props) {
-  //   super(props);
-  // }
-
-  render() {
-    return (
-      <Switch>
-        <Route exact path="/" component={Homepage} />
-        <Route exact path="/signup" component={SignUp} />
-        <Route exact path="/login" component={Login} />
-        <Route exact path="/profile" component={Profile} />
-        <Route exact path="/propertysearch" component={PropertySearch} />
-        <Route exact path="/savedproperties" component={UserSavedProperties} />
-        <Route exact path="/request" component={InfoRequestPage} />
-      </Switch>
-    );
-  }
-}
+const Routes = () => (
+  <Switch>
+    <Route exact path="/" component={Homepage} />
+    <Route exact path="/signup" component={SignUp} />
+    <Route exact path="/login" component={Login} />
+    <Route exact path="/profile" component={Profile} />
+    <Route exact path="/propertysearch" component={PropertySearch} />
+    <Route exact path="/savedproperties" component={UserSavedProperties} />
+    <Route exact path="/request" component={InfoRequestPage} />
+  </Switch>
+);
 
 export default withRouter(Routes);
